+++ src/components/PlaylistToolBar.vue
@@ -32,7 +32,12 @@
     />
     <q-btn icon="edit" class="icon" @click="editButtonClicked" size="sm" />
     <q-btn icon="delete" class="icon" @click="deleteButtonClicked" size="sm" />
-    <q-btn :class="getFadeButtonClass()" @click="fadeButtonClicked" size="sm">
+    <q-btn
+      :class="getFadeButtonClass()"
+      @click="fadeButtonClicked"
+      size="sm"
+      :disable="!fadeButtonEnabled()"
+    >
       <svg
         width="24"
         height="24"
@@ -141,9 +146,23 @@ function deleteButtonClicked() {
 
 function getSvgPath() {
   if (soundsStore.selectedSound === null) return getFadeInSvgPath();
-  if (soundsStore.selectedSound.isPlaying) return getFadeOutSvgPath();
+  if (soundsStore.selectedSound.isPlaying) {
+    if (soundsStore.selectedSound.isFadingIn) return getFadeInSvgPath();
+    return getFadeOutSvgPath();
+  }
   return getFadeInSvgPath();
 }
+
+function fadeButtonEnabled() {
+  if (soundsStore.selectedSound === null) return false;
+  if (
+    soundsStore.selectedSound.isFadingIn ||
+    soundsStore.selectedSound.isFadingOut
+  )
+    return false;
+  return true;
+}
+
 function getFadeInSvgPath() {
   return 'M36 208c-2.21 0-2.619-1.144-.926-2.546L220.926 51.546c1.698-1.406 3.074-.76 3.074 1.464v150.98a4.003 4.003 0 0 1-3.996 4.01h-8.008a3.996 3.996 0 0 1-3.996-4.007V84.007c0-2.213-1.387-2.861-3.079-1.465L56.08 205.458C54.379 206.862 51.209 208 49 208H35.999z';
 }